Move existing application inside the 'use' command
